+++ src/app/modules/select-records-page/select-records-page.component.ts
@@ -14,11 +14,12 @@ import { ErrorStateMatcher } from '@angular/material/core';
 import { LiveAnnouncer } from '@angular/cdk/a11y';
 import { CohortService } from '../../shared/cohort/cohort.service';
 import { Criteria, ResourceTypeCriteria } from '../../types/search-parameters';
-import { SearchParameterGroupComponent } from '../search-parameter-group/search-parameter-group.component';
 import { HttpContext } from '@angular/common/http';
 import { BrowseRecordsPageComponent } from '../browse-records-page/browse-records-page.component';
 import { MatTabGroup } from '@angular/material/tabs';
-import { SearchParameterGroup } from '../../types/search-parameter-group';
+import {
+  SearchParametersComponent
+} from '../search-parameters/search-parameters.component';
 
 /**
  * Component for searching, selecting, and adding records to the cart.
@@ -39,7 +40,7 @@ export class SelectRecordsPageComponent
   extends BrowseRecordsPageComponent
   implements AfterViewInit, OnDestroy {
   @ViewChild('additionalCriteria')
-  additionalCriteria: SearchParameterGroupComponent;
+  additionalCriteria: SearchParametersComponent;
   @ViewChild('tabGroup') tabGroup: MatTabGroup;
   maxPatientsNumber = new UntypedFormControl(
     this.cohort.maxPatientCount,
@@ -294,13 +295,7 @@ export class SelectRecordsPageComponent
    * Gets the cohort criteria from controls.
    */
   getCriteriaFromControls(): Criteria {
-    const additionalCriteria: ResourceTypeCriteria = {
-      condition: 'and',
-      resourceType: 'Patient',
-      rules: this.additionalCriteria
-        .getSearchParamValues()
-        .map((v) => ({field: v}))
-    };
+    const additionalCriteria = this.additionalCriteria.queryCtrl.value;
 
     const variableCriteria: Criteria = this.getVariableCriteria();
 
@@ -340,7 +335,7 @@ export class SelectRecordsPageComponent
    */
   setCartCriteria(
     cartCriteria: any,
-    additionalCriteria: SearchParameterGroup
+    additionalCriteria: any
   ): void {
     // Select the 'Variables' tab so the variable constraint controls is visible.
     // Otherwise, the Prefetch autocomplete of a 'unit' control in 'Variables' cart cannot be setup.
@@ -350,7 +345,7 @@ export class SelectRecordsPageComponent
       // Select and restore 'additional criteria' tab.
       this.tabGroup.selectedIndex = 2;
       setTimeout(() => {
-        this.additionalCriteria.writeValue(additionalCriteria);
+        this.additionalCriteria.queryCtrl.setValue(additionalCriteria);
         setTimeout(() => {
           // Switch back to 'Studies' tab.
           this.tabGroup.selectedIndex = 0;