+++ CHANGELOG.md
@@ -4,6 +4,8 @@ This log documents the significant changes for each release.
 This project follows [Semantic Versioning](http://semver.org/).
 
 ## [6.8.0] - 2023-08-31
+### Added
+- Using the "advanced search" UI for additional criteria.
 ### Fixed
 - A bug in the search algorithm: ANDed criteria may be ORed in the parent node.
 - Issues with saving and loading a cohort and criteria.